requirements.txt
dash==2.16.1
dash==2.7.1
dash==1.17.0
matplotlib==3.8.4
matplotlib==2.2.5
matplotlib==3.3.3
numpy==1.23.5
numpy==1.19.4
plotly==5.20.0
plotly==5.22.0
plotly==4.12.0
pygame==2.5.2
pyzmq==25.1.2
pyzmq==20.0.0
~lotly==4.12.0
